refactor(dis-pgsql): split logical database postgres helpers
